default.json: Include current and previous journal
The system logs were not gathered by default.
Add the output of journalctl for the current and previous boots, but only
for system messages like the dmesg output and logs of system services
(not user sessions). The journal text output can be large but usually
compresses very well. If there was no previous boot, the dump file in
the tar ball will be empty.